+++ docs/changelog.md
@@ -8,6 +8,7 @@
 
 * **Changed:**
     * The Pit Kiln's efficient tool has been changed from axe to shovel [(#412)](https://github.com/codetaylor/pyrotech-1.12/issues/412)
+    * The Campfire's efficient tool has been changed from axe to shovel
 
 * **Added:**
     * Spanish (es_es) localization [(#450)](https://github.com/codetaylor/pyrotech-1.12/pull/450)